benchmark/diffusion/text-to-video/requirements.txt
torch
diffusers==0.29.2
accelerate
datasets
transformers
pillow
nvidia-ml-py
zeus-ml
